FTPD_client: fix compilation warning when compiling with clang
* reply(): change return type from `int` to `ssize_t`
* reply_raw(): change return type from `int` to `ssize_t`

+++ FTPD_client.hh
@@ -40,8 +40,8 @@ public:
 	~FTPD_client();
 
 	void reset();
-	int reply();
-	int reply_raw(int code, const char* msg, const char* msg_add);
+	ssize_t reply();
+	ssize_t reply_raw(int code, const char* msg, const char* msg_add);
 
 	int		_s;
 	int		_conn_stat;
